refatocr: 3d opts
