+++ opts/gl_3d.go
@@ -1,0 +1,175 @@
+package opts
+
+import "github.com/go-echarts/go-echarts/v2/types"
+
+// XAxis3D contains options for X axis in the 3D coordinate.
+type XAxis3D struct {
+	// Whether to display the x-axis.
+	Show types.Bool `json:"show,omitempty"`
+
+	// The name of the axis.
+	Name string `json:"name,omitempty"`
+
+	// The index of the grid3D component used by the axis. The default is to use the first grid3D component.
+	Grid3DIndex int `json:"grid3DIndex,omitempty"`
+
+	// The type of the axis.
+	// Optional:
+	// * 'value' The value axis. Suitable for continuous data.
+	// * 'category' The category axis. Suitable for the discrete category data.
+	//  For this type, the category data must be set through data.
+	// * 'time' The timeline. Suitable for the continuous timing data. The time axis has a
+	//  time format compared to the value axis, and the scale calculation is also different.
+	//  For example, the scale of the month, week, day, and hour ranges can be determined according to the range of the span.
+	// * 'log' Logarithmic axis. Suitable for the logarithmic data.
+	Type string `json:"type,omitempty"`
+
+	// The minimum value of axis.
+	// It can be set to a special value 'dataMin' so that the minimum value on this axis is set to be the minimum label.
+	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
+	// In the category axis, it can also be set as the ordinal number. For example,
+	// if a category axis has data: ['categoryA', 'categoryB', 'categoryC'],
+	// and the ordinal 2 represents 'categoryC'. Moreover, it can be set as a negative number, like -3.
+	Min interface{} `json:"min,omitempty"`
+
+	// The maximum value of the axis.
+	// It can be set to a special value 'dataMax' so that the minimum value on this axis is set to be the maximum label.
+	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
+	// In the category axis, it can also be set as the ordinal number. For example, if a category axis
+	// has data: ['categoryA', 'categoryB', 'categoryC'], and the ordinal 2 represents 'categoryC'.
+	// Moreover, it can be set as a negative number, like -3.
+	Max interface{} `json:"max,omitempty"`
+
+	// Category data, available in type: 'category' axis.
+	// If type is specified as 'category', but axis.data is not specified, axis.data will be auto
+	// collected from series.data. It brings convenience, but we should notice that axis.data provides
+	// then value range of the 'category' axis. If it is auto collected from series.data,
+	// Only the values appearing in series.data can be collected. For example,
+	// if series.data is empty, nothing will be collected.
+	Data interface{} `json:"data,omitempty"`
+}
+
+// YAxis3D contains options for Y axis in the 3D coordinate.
+type YAxis3D struct {
+	// Whether to display the y-axis.
+	Show types.Bool `json:"show,omitempty"`
+
+	// The name of the axis.
+	Name string `json:"name,omitempty"`
+
+	// The index of the grid3D component used by the axis. The default is to use the first grid3D component.
+	Grid3DIndex int `json:"grid3DIndex,omitempty"`
+
+	// The type of the axis.
+	// Optional:
+	// * 'value' The value axis. Suitable for continuous data.
+	// * 'category' The category axis. Suitable for the discrete category data.
+	//  For this type, the category data must be set through data.
+	// * 'time' The timeline. Suitable for the continuous timing data. The time axis has a
+	//  time format compared to the value axis, and the scale calculation is also different.
+	//  For example, the scale of the month, week, day, and hour ranges can be determined according to the range of the span.
+	// * 'log' Logarithmic axis. Suitable for the logarithmic data.
+	Type string `json:"type,omitempty"`
+
+	// The minimum value of axis.
+	// It can be set to a special value 'dataMin' so that the minimum value on this axis is set to be the minimum label.
+	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
+	// In the category axis, it can also be set as the ordinal number. For example,
+	// if a category axis has data: ['categoryA', 'categoryB', 'categoryC'],
+	// and the ordinal 2 represents 'categoryC'. Moreover, it can be set as a negative number, like -3.
+	Min interface{} `json:"min,omitempty"`
+
+	// The maximum value of the axis.
+	// It can be set to a special value 'dataMax' so that the minimum value on this axis is set to be the maximum label.
+	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
+	// In the category axis, it can also be set as the ordinal number. For example, if a category axis
+	// has data: ['categoryA', 'categoryB', 'categoryC'], and the ordinal 2 represents 'categoryC'.
+	// Moreover, it can be set as a negative number, like -3.
+	Max interface{} `json:"max,omitempty"`
+
+	// Category data, available in type: 'category' axis.
+	// If type is specified as 'category', but axis.data is not specified, axis.data will be auto
+	// collected from series.data. It brings convenience, but we should notice that axis.data provides
+	// then value range of the 'category' axis. If it is auto collected from series.data,
+	// Only the values appearing in series.data can be collected. For example,
+	// if series.data is empty, nothing will be collected.
+	Data interface{} `json:"data,omitempty"`
+}
+
+// ZAxis3D contains options for Z axis in the 3D coordinate.
+type ZAxis3D struct {
+	// Whether to display the z-axis.
+	Show types.Bool `json:"show,omitempty"`
+
+	// The name of the axis.
+	Name string `json:"name,omitempty"`
+
+	// The index of the grid3D component used by the axis. The default is to use the first grid3D component.
+	Grid3DIndex int `json:"grid3DIndex,omitempty"`
+
+	// The type of the axis.
+	// Optional:
+	// * 'value' The value axis. Suitable for continuous data.
+	// * 'category' The category axis. Suitable for the discrete category data.
+	//  For this type, the category data must be set through data.
+	// * 'time' The timeline. Suitable for the continuous timing data. The time axis has a
+	//  time format compared to the value axis, and the scale calculation is also different.
+	//  For example, the scale of the month, week, day, and hour ranges can be determined according to the range of the span.
+	// * 'log' Logarithmic axis. Suitable for the logarithmic data.
+	Type string `json:"type,omitempty"`
+
+	// The minimum value of axis.
+	// It can be set to a special value 'dataMin' so that the minimum value on this axis is set to be the minimum label.
+	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
+	// In the category axis, it can also be set as the ordinal number. For example,
+	// if a category axis has data: ['categoryA', 'categoryB', 'categoryC'],
+	// and the ordinal 2 represents 'categoryC'. Moreover, it can be set as a negative number, like -3.
+	Min interface{} `json:"min,omitempty"`
+
+	// The maximum value of the axis.
+	// It can be set to a special value 'dataMax' so that the minimum value on this axis is set to be the maximum label.
+	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
+	// In the category axis, it can also be set as the ordinal number. For example, if a category axis
+	// has data: ['categoryA', 'categoryB', 'categoryC'], and the ordinal 2 represents 'categoryC'.
+	// Moreover, it can be set as a negative number, like -3.
+	Max interface{} `json:"max,omitempty"`
+
+	// Category data, available in type: 'category' axis.
+	// If type is specified as 'category', but axis.data is not specified, axis.data will be auto
+	// collected from series.data. It brings convenience, but we should notice that axis.data provides
+	// then value range of the 'category' axis. If it is auto collected from series.data,
+	// Only the values appearing in series.data can be collected. For example,
+	// if series.data is empty, nothing will be collected.
+	Data interface{} `json:"data,omitempty"`
+}
+
+// Grid3D contains options for the 3D coordinate.
+type Grid3D struct {
+	// Whether to show the coordinate.
+	Show types.Bool `json:"show,omitempty"`
+
+	// 3D Cartesian coordinates width
+	// default 100
+	BoxWidth float32 `json:"boxWidth,omitempty"`
+
+	// 3D Cartesian coordinates height
+	// default 100
+	BoxHeight float32 `json:"boxHeight,omitempty"`
+
+	// 3D Cartesian coordinates depth
+	// default 100
+	BoxDepth float32 `json:"boxDepth,omitempty"`
+
+	// Rotate or scale fellows the mouse
+	ViewControl *ViewControl `json:"viewControl,omitempty"`
+}
+
+// ViewControl contains options for view controlling.
+type ViewControl struct {
+	// Whether to rotate automatically.
+	AutoRotate types.Bool `json:"autoRotate,omitempty"`
+
+	// Rotate Speed, (angle/s).
+	// default 10
+	AutoRotateSpeed float32 `json:"autoRotateSpeed,omitempty"`
+}
