+++ opts/global.go
@@ -110,130 +110,6 @@ func randByte() byte {
 	return byte(c + rand.Intn(26))
 }
 
-// AxisLabel settings related to axis label.
-// https://echarts.apache.org/en/option.html#xAxis.axisLabel
-type AxisLabel struct {
-	// Set this to false to prevent the axis label from appearing.
-	Show types.Bool `json:"show,omitempty"`
-
-	// Interval of Axis label, which is available in category axis.
-	// It uses a strategy that labels do not overlap by default.
-	// You may set it to be 0 to display all labels compulsively.
-	// If it is set to be 1, it means that labels are shown once after one label.
-	// And if it is set to be 2, it means labels are shown once after two labels, and so on.
-	Interval string `json:"interval,omitempty"`
-
-	// Set this to true so the axis labels face the inside direction.
-	Inside types.Bool `json:"inside,omitempty"`
-
-	// Rotation degree of axis label, which is especially useful when there is no enough space for category axis.
-	// Rotation degree is from -90 to 90.
-	Rotate float64 `json:"rotate,omitempty"`
-
-	// The margin between the axis label and the axis line.
-	Margin float64 `json:"margin,omitempty"`
-
-	// Formatter of axis label, which supports string template and callback function.
-	//
-	// Example:
-	//
-	// Use string template; template variable is the default label of axis {value}
-	// formatter: '{value} kg'
-	//
-	// Use callback function; function parameters are axis index
-	//
-	//
-	//  formatter: function (value, index) {
-	//    // Formatted to be month/day; display year only in the first label
-	//    var date = new Date(value);
-	//    var texts = [(date.getMonth() + 1), date.getDate()];
-	//    if (idx === 0) {
-	//        texts.unshift(date.getYear());
-	//    }
-	//    return texts.join('/');
-	// }
-	Formatter string `json:"formatter,omitempty"`
-
-	ShowMinLabel types.Bool `json:"showMinLabel"`
-	ShowMaxLabel types.Bool `json:"showMaxLabel"`
-
-	// Color of axis label is set to be axisLine.lineStyle.color by default. Callback function is supported,
-	// in the following format:
-	//
-	// (val: string) => Color
-	// Parameter is the text of label, and return value is the color. See the following example:
-	//
-	// textStyle: {
-	//    color: function (value, index) {
-	//        return value >= 0 ? 'green' : 'red';
-	//    }
-	// }
-	Color string `json:"color,omitempty"`
-
-	// axis label font style
-	FontStyle string `json:"fontStyle,omitempty"`
-	// axis label font weight
-	FontWeight string `json:"fontWeight,omitempty"`
-	// axis label font family
-	FontFamily string `json:"fontFamily,omitempty"`
-	// axis label font size
-	FontSize string `json:"fontSize,omitempty"`
-	// Horizontal alignment of axis label
-	Align string `json:"align,omitempty"`
-	// Vertical alignment of axis label
-	VerticalAlign string `json:"verticalAlign,omitempty"`
-	// Line height of the axis label
-	LineHeight string `json:"lineHeight,omitempty"`
-
-	BackgroundColor string `json:"backgroundColor,omitempty"`
-}
-
-type AxisTick struct {
-	// Set this to false to prevent the axis tick from showing.
-	Show types.Bool `json:"show,omitempty"`
-
-	// interval of axisTick, which is available in category axis. is set to be the same as axisLabel.interval by default.
-	// It uses a strategy that labels do not overlap by default.
-	// You may set it to be 0 to display all labels compulsively.
-	// If it is set to be 1, it means that labels are shown once after one label. And if it is set to be 2, it means labels are shown once after two labels, and so on.
-	// On the other hand, you can control by callback function, whose format is shown below:
-	// (index:number, value: string) => types.Boolean
-	// The first parameter is index of category, and the second parameter is the name of category. The return values decides whether to display label.
-	Interval string `json:"interval,omitempty"`
-
-	// Align axis tick with label, which is available only when boundaryGap is set to be true in category axis.
-	AlignWithLabel types.Bool `json:"alignWithLabel,omitempty"`
-}
-
-// AxisLine controls settings related to axis line.
-// https://echarts.apache.org/en/option.html#yAxis.axisLine
-type AxisLine struct {
-	// Set this to false to prevent the axis line from showing.
-	Show types.Bool `json:"show,omitempty"`
-
-	// Specifies whether X or Y axis lies on the other's origin position, where value is 0 on axis.
-	// Valid only if the other axis is of value type, and contains 0 value.
-	OnZero types.Bool `json:"onZero,omitempty"`
-
-	// When multiple axes exists, this option can be used to specify which axis can be "onZero" to.
-	OnZeroAxisIndex int `json:"onZeroAxisIndex,omitempty"`
-
-	// Symbol of the two ends of the axis. It could be a string, representing the same symbol for two ends; or an array
-	// with two string elements, representing the two ends separately. It's set to be 'none' by default, meaning no
-	//arrow for either end. If it is set to be 'arrow', there shall be two arrows. If there should only one arrow
-	//at the end, it should set to be ['none', 'arrow'].
-	Symbol string `json:"symbol,omitempty"`
-
-	// Size of the arrows at two ends. The first is the width perpendicular to the axis, the next is the width parallel to the axis.
-	SymbolSize []float64 `json:"symbolSize,omitempty"`
-
-	// Arrow offset of axis. If is array, the first number is the offset of the arrow at the beginning, and the second
-	// number is the offset of the arrow at the end. If is number, it means the arrows have the same offset.
-	SymbolOffset []float64 `json:"symbolOffset,omitempty"`
-
-	LineStyle *LineStyle `json:"lineStyle,omitempty"`
-}
-
 // SplitArea is the option set for a split area.
 type SplitArea struct {
 	// Set this to true to show the splitArea.
@@ -338,176 +214,4 @@ func (opt *Assets) Validate(host string) {
 			opt.CSSAssets.Values[i] = host + opt.CSSAssets.Values[i]
 		}
 	}
-}
-
-// XAxis3D contains options for X axis in the 3D coordinate.
-type XAxis3D struct {
-	// Whether to display the x-axis.
-	Show types.Bool `json:"show,omitempty"`
-
-	// The name of the axis.
-	Name string `json:"name,omitempty"`
-
-	// The index of the grid3D component used by the axis. The default is to use the first grid3D component.
-	Grid3DIndex int `json:"grid3DIndex,omitempty"`
-
-	// The type of the axis.
-	// Optional:
-	// * 'value' The value axis. Suitable for continuous data.
-	// * 'category' The category axis. Suitable for the discrete category data.
-	//  For this type, the category data must be set through data.
-	// * 'time' The timeline. Suitable for the continuous timing data. The time axis has a
-	//  time format compared to the value axis, and the scale calculation is also different.
-	//  For example, the scale of the month, week, day, and hour ranges can be determined according to the range of the span.
-	// * 'log' Logarithmic axis. Suitable for the logarithmic data.
-	Type string `json:"type,omitempty"`
-
-	// The minimum value of axis.
-	// It can be set to a special value 'dataMin' so that the minimum value on this axis is set to be the minimum label.
-	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
-	// In the category axis, it can also be set as the ordinal number. For example,
-	// if a category axis has data: ['categoryA', 'categoryB', 'categoryC'],
-	// and the ordinal 2 represents 'categoryC'. Moreover, it can be set as a negative number, like -3.
-	Min interface{} `json:"min,omitempty"`
-
-	// The maximum value of the axis.
-	// It can be set to a special value 'dataMax' so that the minimum value on this axis is set to be the maximum label.
-	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
-	// In the category axis, it can also be set as the ordinal number. For example, if a category axis
-	// has data: ['categoryA', 'categoryB', 'categoryC'], and the ordinal 2 represents 'categoryC'.
-	// Moreover, it can be set as a negative number, like -3.
-	Max interface{} `json:"max,omitempty"`
-
-	// Category data, available in type: 'category' axis.
-	// If type is specified as 'category', but axis.data is not specified, axis.data will be auto
-	// collected from series.data. It brings convenience, but we should notice that axis.data provides
-	// then value range of the 'category' axis. If it is auto collected from series.data,
-	// Only the values appearing in series.data can be collected. For example,
-	// if series.data is empty, nothing will be collected.
-	Data interface{} `json:"data,omitempty"`
-}
-
-// YAxis3D contains options for Y axis in the 3D coordinate.
-type YAxis3D struct {
-	// Whether to display the y-axis.
-	Show types.Bool `json:"show,omitempty"`
-
-	// The name of the axis.
-	Name string `json:"name,omitempty"`
-
-	// The index of the grid3D component used by the axis. The default is to use the first grid3D component.
-	Grid3DIndex int `json:"grid3DIndex,omitempty"`
-
-	// The type of the axis.
-	// Optional:
-	// * 'value' The value axis. Suitable for continuous data.
-	// * 'category' The category axis. Suitable for the discrete category data.
-	//  For this type, the category data must be set through data.
-	// * 'time' The timeline. Suitable for the continuous timing data. The time axis has a
-	//  time format compared to the value axis, and the scale calculation is also different.
-	//  For example, the scale of the month, week, day, and hour ranges can be determined according to the range of the span.
-	// * 'log' Logarithmic axis. Suitable for the logarithmic data.
-	Type string `json:"type,omitempty"`
-
-	// The minimum value of axis.
-	// It can be set to a special value 'dataMin' so that the minimum value on this axis is set to be the minimum label.
-	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
-	// In the category axis, it can also be set as the ordinal number. For example,
-	// if a category axis has data: ['categoryA', 'categoryB', 'categoryC'],
-	// and the ordinal 2 represents 'categoryC'. Moreover, it can be set as a negative number, like -3.
-	Min interface{} `json:"min,omitempty"`
-
-	// The maximum value of the axis.
-	// It can be set to a special value 'dataMax' so that the minimum value on this axis is set to be the maximum label.
-	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
-	// In the category axis, it can also be set as the ordinal number. For example, if a category axis
-	// has data: ['categoryA', 'categoryB', 'categoryC'], and the ordinal 2 represents 'categoryC'.
-	// Moreover, it can be set as a negative number, like -3.
-	Max interface{} `json:"max,omitempty"`
-
-	// Category data, available in type: 'category' axis.
-	// If type is specified as 'category', but axis.data is not specified, axis.data will be auto
-	// collected from series.data. It brings convenience, but we should notice that axis.data provides
-	// then value range of the 'category' axis. If it is auto collected from series.data,
-	// Only the values appearing in series.data can be collected. For example,
-	// if series.data is empty, nothing will be collected.
-	Data interface{} `json:"data,omitempty"`
-}
-
-// ZAxis3D contains options for Z axis in the 3D coordinate.
-type ZAxis3D struct {
-	// Whether to display the z-axis.
-	Show types.Bool `json:"show,omitempty"`
-
-	// The name of the axis.
-	Name string `json:"name,omitempty"`
-
-	// The index of the grid3D component used by the axis. The default is to use the first grid3D component.
-	Grid3DIndex int `json:"grid3DIndex,omitempty"`
-
-	// The type of the axis.
-	// Optional:
-	// * 'value' The value axis. Suitable for continuous data.
-	// * 'category' The category axis. Suitable for the discrete category data.
-	//  For this type, the category data must be set through data.
-	// * 'time' The timeline. Suitable for the continuous timing data. The time axis has a
-	//  time format compared to the value axis, and the scale calculation is also different.
-	//  For example, the scale of the month, week, day, and hour ranges can be determined according to the range of the span.
-	// * 'log' Logarithmic axis. Suitable for the logarithmic data.
-	Type string `json:"type,omitempty"`
-
-	// The minimum value of axis.
-	// It can be set to a special value 'dataMin' so that the minimum value on this axis is set to be the minimum label.
-	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
-	// In the category axis, it can also be set as the ordinal number. For example,
-	// if a category axis has data: ['categoryA', 'categoryB', 'categoryC'],
-	// and the ordinal 2 represents 'categoryC'. Moreover, it can be set as a negative number, like -3.
-	Min interface{} `json:"min,omitempty"`
-
-	// The maximum value of the axis.
-	// It can be set to a special value 'dataMax' so that the minimum value on this axis is set to be the maximum label.
-	// It will be automatically computed to make sure the axis tick is equally distributed when not set.
-	// In the category axis, it can also be set as the ordinal number. For example, if a category axis
-	// has data: ['categoryA', 'categoryB', 'categoryC'], and the ordinal 2 represents 'categoryC'.
-	// Moreover, it can be set as a negative number, like -3.
-	Max interface{} `json:"max,omitempty"`
-
-	// Category data, available in type: 'category' axis.
-	// If type is specified as 'category', but axis.data is not specified, axis.data will be auto
-	// collected from series.data. It brings convenience, but we should notice that axis.data provides
-	// then value range of the 'category' axis. If it is auto collected from series.data,
-	// Only the values appearing in series.data can be collected. For example,
-	// if series.data is empty, nothing will be collected.
-	Data interface{} `json:"data,omitempty"`
-}
-
-// Grid3D contains options for the 3D coordinate.
-type Grid3D struct {
-	// Whether to show the coordinate.
-	Show types.Bool `json:"show,omitempty"`
-
-	// 3D Cartesian coordinates width
-	// default 100
-	BoxWidth float32 `json:"boxWidth,omitempty"`
-
-	// 3D Cartesian coordinates height
-	// default 100
-	BoxHeight float32 `json:"boxHeight,omitempty"`
-
-	// 3D Cartesian coordinates depth
-	// default 100
-	BoxDepth float32 `json:"boxDepth,omitempty"`
-
-	// Rotate or scale fellows the mouse
-	ViewControl *ViewControl `json:"viewControl,omitempty"`
-}
-
-// ViewControl contains options for view controlling.
-type ViewControl struct {
-	// Whether to rotate automatically.
-	AutoRotate types.Bool `json:"autoRotate,omitempty"`
-
-	// Rotate Speed, (angle/s).
-	// default 10
-	AutoRotateSpeed float32 `json:"autoRotateSpeed,omitempty"`
 }
